use factory pattern createReactRouterRoutingProvider
