+++ packages/platforms/browser/lib/browser.ts
@@ -3,7 +3,7 @@ import { FullPageLoadPlugin, NetworkRequestPlugin, ResourceLoadPlugin, RouteChan
 import createBrowserBackgroundingListener from './backgrounding-listener'
 import createClock from './clock'
 import { createSchema } from './config'
-import { createDefaultRoutingProvider } from './default-routing-provider'
+import { createDefaultRoutingProvider as _createDefaultRoutingProvider } from './default-routing-provider'
 import createBrowserDeliveryFactory from './delivery'
 import idGenerator from './id-generator'
 import createOnSettle from './on-settle'
@@ -29,7 +29,9 @@ export const onSettle = createOnSettle(
   xhrRequestTracker,
   performance
 )
-export const DefaultRoutingProvider = createDefaultRoutingProvider(onSettle, window.location)
+export const createDefaultRoutingProvider = _createDefaultRoutingProvider
+export const DefaultRoutingProvider = _createDefaultRoutingProvider(onSettle, window.location)
+export { type OnSettle } from './on-settle'
 
 const BugsnagPerformance = createClient({
   backgroundingListener,
