+++ packages/react-router/lib/index.ts
@@ -1,27 +1,4 @@
-import { DefaultRoutingProvider } from '@bugsnag/browser-performance'
-import { matchPath } from 'react-router-dom'
+import { onSettle } from '@bugsnag/browser-performance'
+import { createReactRouterRoutingProvider } from './react-router-routing-provider'
 
-export interface RouteObject {
-  path?: string
-  children?: RouteObject[]
-}
-
-function flattenRoutes (routes: RouteObject[], _prefix: string = ''): string[] {
-  const prefix = `${!_prefix || _prefix === '/' ? _prefix : `${_prefix}/`}`
-  return [
-    ...routes.map(route => `${prefix}${route.path || ''}`),
-    ...routes.reduce<string[]>(
-      (accum, route) => [...accum, ...(route.children ? flattenRoutes(route.children, `${prefix}${route.path}`) : [])],
-      []
-    )
-  ]
-}
-
-export class ReactRouterRoutingProvider extends DefaultRoutingProvider {
-  constructor (routes: RouteObject[], basename?: string) {
-    function resolveRoute (url: URL): string {
-      return flattenRoutes(routes).find((fullRoutePath) => matchPath(fullRoutePath, url.pathname.replace(basename ?? '', ''))?.pattern.path) || 'no-route-found'
-    }
-    super(resolveRoute)
-  }
-}
+export const ReactRouterRoutingProvider = createReactRouterRoutingProvider(onSettle, window.location)
