+++ packages/react-router/tests/react-router-routing-provider.test.ts
@@ -2,8 +2,7 @@
  * @jest-environment jsdom
  */
 
-import { type Span, SpanAttributes, SpanInternal } from '@bugsnag/core-performance'
-import { ReactRouterRoutingProvider } from '../lib'
+import { createReactRouterRoutingProvider } from '../lib/react-router-routing-provider'
 
 jest.useFakeTimers()
 
@@ -29,10 +28,11 @@ describe('ReactRouterRoutingProvider', () => {
 
   describe('listenForRouteChanges', () => {
     it('invokes the provided callback on pushState', () => {
+      const onSettle = jest.fn(() => 1234)
+      const ReactRouterRoutingProvider = createReactRouterRoutingProvider(onSettle, window.location)
+
       const routingProvider = new ReactRouterRoutingProvider(routes)
-      const spanAttributes = new SpanAttributes(new Map())
-      const span = new SpanInternal('id', 'traceId', 'test span', 1234, spanAttributes) as unknown as Span
-      const startRouteChangeSpan = jest.fn(() => span)
+      const startRouteChangeSpan = jest.fn()
 
       routingProvider.listenForRouteChanges(startRouteChangeSpan)
 
@@ -42,6 +42,9 @@ describe('ReactRouterRoutingProvider', () => {
     })
 
     it('invokes the provided callback on popstate', () => {
+      const onSettle = jest.fn(() => 1234)
+      const ReactRouterRoutingProvider = createReactRouterRoutingProvider(onSettle, window.location)
+
       const routingProvider = new ReactRouterRoutingProvider(routes)
       const startRouteChangeSpan = jest.fn()
 
@@ -59,6 +62,9 @@ describe('ReactRouterRoutingProvider', () => {
 
   describe('resolveRoute', () => {
     it('uses the provided routes when resolving routes', () => {
+      const onSettle = jest.fn(() => 1234)
+      const ReactRouterRoutingProvider = createReactRouterRoutingProvider(onSettle, window.location)
+
       const routingProvider = new ReactRouterRoutingProvider(routes)
 
       expect(routingProvider.resolveRoute(new URL(window.location.href))).toBe('/contacts/:contactId')
@@ -71,6 +77,9 @@ describe('ReactRouterRoutingProvider', () => {
 
   describe('basename', () => {
     it('uses the provided basename when resolving routes', () => {
+      const onSettle = jest.fn(() => 1234)
+      const ReactRouterRoutingProvider = createReactRouterRoutingProvider(onSettle, window.location)
+
       const routingProvider = new ReactRouterRoutingProvider(routes, '/app')
 
       history.pushState({}, '', '/app/contacts/2/edit')
